+++ Project.toml
@@ -4,7 +4,6 @@ authors = ["SimonDanisch <sdanisch@gmail.com>"]
 version = "0.6.13"
 
 [deps]
-AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
 Base64 = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
@@ -22,11 +21,10 @@ UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 WidgetsBase = "eead4739-05f7-45a1-878c-cee36b57321c"
 
 [compat]
-AbstractTrees = "0.3"
 Colors = "0.12"
 HTTP = "0.8, 0.9"
 Hyperscript = "0.0.3, 0.0.4"
-JSON3 = "0.1, 1.0"
+JSON3 = "1.0"
 MsgPack = "1.1.0"
 Observables = "0.3"
 Tables = "1"
